refactor: サインアップ処理のメソッド切り出し.
