+++ lib/constant.dart
@@ -3,6 +3,7 @@ class AssetPath {
   static const GOOGLE_LOGO_PATH = 'assets/img/ic_logo_google.png';
   static const TW_LOGO_PATH = 'assets/img/ic_logo_twitter.png';
   static const MAIL_LOGO_PATH = 'assets/img/ic_logo_mail.png';
+  static const PROFILE_DUMMY_IMG = 'assets/img/profile_dummy.png';
 }
 
 class Pref {
