+++ lib/repositories/auth_repository.dart
@@ -3,7 +3,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/services.dart';
 import 'package:nippo/models/user.dart';
 import 'package:google_sign_in/google_sign_in.dart';
-import 'package:nippo/dateutil.dart';
+import 'package:nippo/util.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:nippo/constant.dart';
